Scaffold fresh Rails-API app
After diving deeper into Rails-built microservices I decided to test out the new API-first features merged into core Rails in version 5. This functionality previously existed as the `Rails-API` gem.

This should make it easier to create a RESTful interface for other applications, and will eliminate the scaffolding overhead that comes with the `views/` and `assets/` directories.