Scaffold the submission model
This is the early stages of handling a form submission and just focuses on a simple request/response.

This version saves submissions to a Postgres database, which I thought might be valuable for the sake of the city's policy on record-keeping. This functionality, however, could just as easily be federated out to the dispatched microservices, though in that case our ability to diagnose dispatcher failures would be more constrained to the data available in server logs (which are not currently being kept).