Install CORS
In order for the dispatcher to run independently it must be able to accept requests many possible domains. The rack-cors gem allows us to whitelist domains from which we want to accept and process requests.

In my local development I've arbitrarily decided to run the client-side testing part of this on port 3001, which is why that's the declared value for the development environment.

The heroku production url is where the sandbox form sandbox site is hosted.